requirements.txt
torch>=2.0.0
transformers>=4.37.0
peft>=0.7.0
accelerate>=0.25.0
Pillow
python-dotenv
tqdm
wandb